chore: relocate application runner to own file
